+++ src/demo-app/main.ts
@@ -26,4 +26,5 @@ export function main(): Promise<any> {
     .catch((err) => console.error(err));
 }
 
+
 bootloader(main);